+++ src/output/exports.ts
@@ -28,8 +28,6 @@ export function stringifyExport(name: string, type: TypeModel) {
       return `${stringifyComment(type)}export ${type.const ? 'const enum' : 'enum'} ${name} ${stringifyEnum(
         type.values,
       )}`;
-    case 'default':
-      return `${stringifyComment(type)}export default ${stringifyNode(type.value)};`;
     case 'const':
       return `${stringifyComment(type)}export const ${name}: ${stringifyNode(type.value)};`;
     case 'function':